chore(cloudflare): move custom hostnames logic to dedicated files
Also fix casing for some variable and function names.
