+++ provider/cloudflare/cloudflare.go
@@ -20,10 +20,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"maps"
 	"net/http"
 	"os"
-	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -83,36 +81,6 @@ type DNSRecordIndex struct {
 
 type DNSRecordsMap map[DNSRecordIndex]dns.RecordResponse
 
-// customHostname represents a Cloudflare custom hostname (v5 API compatible wrapper)
-type customHostname struct {
-	id                 string
-	hostname           string
-	customOriginServer string
-	customOriginSNI    string
-	ssl                *customHostnameSSL
-}
-
-// customHostnameSSL represents SSL configuration for custom hostname
-type customHostnameSSL struct {
-	sslType              string
-	method               string
-	bundleMethod         string
-	certificateAuthority string
-	settings             customHostnameSSLSettings
-}
-
-// customHostnameSSLSettings represents SSL settings for custom hostname
-type customHostnameSSLSettings struct {
-	minTLSVersion string
-}
-
-// for faster getcustomHostname() lookup
-type customHostnameIndex struct {
-	hostname string
-}
-
-type customHostnamesMap map[customHostnameIndex]customHostname
-
 var recordTypeProxyNotSupported = map[string]bool{
 	"LOC": true,
 	"MX":  true,
@@ -120,17 +88,6 @@ var recordTypeProxyNotSupported = map[string]bool{
 	"SPF": true,
 	"TXT": true,
 	"SRV": true,
-}
-
-type CustomHostnamesConfig struct {
-	Enabled              bool
-	MinTLSVersion        string
-	CertificateAuthority string
-}
-
-var recordTypeCustomHostnameSupported = map[string]bool{
-	"A":     true,
-	"CNAME": true,
 }
 
 // cloudFlareDNS is the subset of the CloudFlare API that we actually use.  Add methods as required. Signatures must match exactly.
@@ -147,8 +104,8 @@ type cloudFlareDNS interface {
 	UpdateDataLocalizationRegionalHostname(ctx context.Context, hostname string, params addressing.RegionalHostnameEditParams) error
 	DeleteDataLocalizationRegionalHostname(ctx context.Context, hostname string, params addressing.RegionalHostnameDeleteParams) error
 	CustomHostnames(ctx context.Context, zoneID string) autoPager[custom_hostnames.CustomHostnameListResponse]
-	DeletecustomHostname(ctx context.Context, customHostnameID string, params custom_hostnames.CustomHostnameDeleteParams) error
-	CreatecustomHostname(ctx context.Context, zoneID string, ch customHostname) error
+	DeleteCustomHostname(ctx context.Context, customHostnameID string, params custom_hostnames.CustomHostnameDeleteParams) error
+	CreateCustomHostname(ctx context.Context, zoneID string, ch customHostname) error
 }
 
 type zoneService struct {
@@ -198,42 +155,6 @@ func (z zoneService) ListZones(ctx context.Context, params zones.ZoneListParams)
 
 func (z zoneService) GetZone(ctx context.Context, zoneID string) (*zones.Zone, error) {
 	return z.service.Zones.Get(ctx, zones.ZoneGetParams{ZoneID: cloudflare.F(zoneID)})
-}
-
-func (z zoneService) CustomHostnames(ctx context.Context, zoneID string) autoPager[custom_hostnames.CustomHostnameListResponse] {
-	params := custom_hostnames.CustomHostnameListParams{
-		ZoneID: cloudflare.F(zoneID),
-	}
-	return z.service.CustomHostnames.ListAutoPaging(ctx, params)
-}
-
-// listAllCustomHostnames extracts all custom hostnames from the iterator
-func listAllCustomHostnames(iter autoPager[custom_hostnames.CustomHostnameListResponse]) ([]customHostname, error) {
-	var customHostnames []customHostname
-	for ch := range autoPagerIterator(iter) {
-		customHostnames = append(customHostnames, customHostname{
-			id:                 ch.ID,
-			hostname:           ch.Hostname,
-			customOriginServer: ch.CustomOriginServer,
-			customOriginSNI:    ch.CustomOriginSNI,
-		})
-	}
-	if iter.Err() != nil {
-		return nil, iter.Err()
-	}
-	return customHostnames, nil
-}
-
-func (z zoneService) DeletecustomHostname(ctx context.Context, customHostnameID string, params custom_hostnames.CustomHostnameDeleteParams) error {
-	_, err := z.service.CustomHostnames.Delete(ctx, customHostnameID, params)
-	return err
-}
-
-func (z zoneService) CreatecustomHostname(ctx context.Context, zoneID string, ch customHostname) error {
-	params := buildCustomHostnameNewParams(zoneID, ch)
-	_, err := z.service.CustomHostnames.New(ctx, params,
-		option.WithJSONSet("custom_origin_server", ch.customOriginServer))
-	return err
 }
 
 // listZonesV4Params returns the appropriate Zone List Params for v4 API
@@ -370,36 +291,6 @@ func convertCloudflareError(err error) error {
 	}
 
 	return err
-}
-
-// buildCustomHostnameNewParams builds the params for creating a custom hostname
-func buildCustomHostnameNewParams(zoneID string, ch customHostname) custom_hostnames.CustomHostnameNewParams {
-	params := custom_hostnames.CustomHostnameNewParams{
-		ZoneID:   cloudflare.F(zoneID),
-		Hostname: cloudflare.F(ch.hostname),
-	}
-	if ch.ssl != nil {
-		sslParams := custom_hostnames.CustomHostnameNewParamsSSL{}
-		if ch.ssl.method != "" {
-			sslParams.Method = cloudflare.F(custom_hostnames.DCVMethod(ch.ssl.method))
-		}
-		if ch.ssl.sslType != "" {
-			sslParams.Type = cloudflare.F(custom_hostnames.DomainValidationType(ch.ssl.sslType))
-		}
-		if ch.ssl.bundleMethod != "" {
-			sslParams.BundleMethod = cloudflare.F(custom_hostnames.BundleMethod(ch.ssl.bundleMethod))
-		}
-		if ch.ssl.certificateAuthority != "" && ch.ssl.certificateAuthority != "none" {
-			sslParams.CertificateAuthority = cloudflare.F(cloudflare.CertificateCA(ch.ssl.certificateAuthority))
-		}
-		if ch.ssl.settings.minTLSVersion != "" {
-			sslParams.Settings = cloudflare.F(custom_hostnames.CustomHostnameNewParamsSSLSettings{
-				MinTLSVersion: cloudflare.F(custom_hostnames.CustomHostnameNewParamsSSLSettingsMinTLSVersion(ch.ssl.settings.minTLSVersion)),
-			})
-		}
-		params.SSL = cloudflare.F(sslParams)
-	}
-	return params
 }
 
 // NewCloudFlareProvider initializes a new CloudFlare DNS based Provider.
@@ -616,102 +507,6 @@ func (p *CloudFlareProvider) ApplyChanges(ctx context.Context, changes *plan.Cha
 	return p.submitChanges(ctx, cloudflareChanges)
 }
 
-// submitCustomHostnameChanges implements Custom Hostname functionality for the Change, returns false if it fails
-func (p *CloudFlareProvider) submitCustomHostnameChanges(ctx context.Context, zoneID string, change *cloudFlareChange, chs customHostnamesMap, logFields log.Fields) bool {
-	// return early if disabled
-	if !p.CustomHostnamesConfig.Enabled {
-		return true
-	}
-
-	switch change.Action {
-	case cloudFlareUpdate:
-		return p.processCustomHostnameUpdate(ctx, zoneID, change, chs, logFields)
-	case cloudFlareDelete:
-		return p.processCustomHostnameDelete(ctx, zoneID, change, chs, logFields)
-	case cloudFlareCreate:
-		return p.processCustomHostnameCreate(ctx, zoneID, change, chs, logFields)
-	}
-
-	return true
-}
-
-func (p *CloudFlareProvider) processCustomHostnameUpdate(ctx context.Context, zoneID string, change *cloudFlareChange, chs customHostnamesMap, logFields log.Fields) bool {
-	if !recordTypeCustomHostnameSupported[string(change.ResourceRecord.Type)] {
-		return true
-	}
-	failedChange := false
-	add, remove, _ := provider.Difference(change.CustomHostnamesPrev, slices.Collect(maps.Keys(change.CustomHostnames)))
-
-	for _, changeCH := range remove {
-		if prevCh, err := getcustomHostname(chs, changeCH); err == nil {
-			prevChID := prevCh.id
-			if prevChID != "" {
-				log.WithFields(logFields).Infof("Removing previous custom hostname %q/%q", prevChID, changeCH)
-				params := custom_hostnames.CustomHostnameDeleteParams{ZoneID: cloudflare.F(zoneID)}
-				chErr := p.Client.DeletecustomHostname(ctx, prevChID, params)
-				if chErr != nil {
-					failedChange = true
-					log.WithFields(logFields).Errorf("failed to remove previous custom hostname %q/%q: %v", prevChID, changeCH, chErr)
-				}
-			}
-		}
-	}
-	for _, changeCH := range add {
-		log.WithFields(logFields).Infof("Adding custom hostname %q", changeCH)
-		chErr := p.Client.CreatecustomHostname(ctx, zoneID, change.CustomHostnames[changeCH])
-		if chErr != nil {
-			failedChange = true
-			log.WithFields(logFields).Errorf("failed to add custom hostname %q: %v", changeCH, chErr)
-		}
-	}
-	return !failedChange
-}
-
-func (p *CloudFlareProvider) processCustomHostnameDelete(ctx context.Context, zoneID string, change *cloudFlareChange, chs customHostnamesMap, logFields log.Fields) bool {
-	failedChange := false
-	for _, changeCH := range change.CustomHostnames {
-		if recordTypeCustomHostnameSupported[string(change.ResourceRecord.Type)] && changeCH.hostname != "" {
-			log.WithFields(logFields).Infof("Deleting custom hostname %q", changeCH.hostname)
-			if ch, err := getcustomHostname(chs, changeCH.hostname); err == nil {
-				chID := ch.id
-				params := custom_hostnames.CustomHostnameDeleteParams{ZoneID: cloudflare.F(zoneID)}
-				chErr := p.Client.DeletecustomHostname(ctx, chID, params)
-				if chErr != nil {
-					failedChange = true
-					log.WithFields(logFields).Errorf("failed to delete custom hostname %q/%q: %v", chID, changeCH.hostname, chErr)
-				}
-			} else {
-				log.WithFields(logFields).Warnf("failed to delete custom hostname %q: %v", changeCH.hostname, err)
-			}
-		}
-	}
-	return !failedChange
-}
-
-func (p *CloudFlareProvider) processCustomHostnameCreate(ctx context.Context, zoneID string, change *cloudFlareChange, chs customHostnamesMap, logFields log.Fields) bool {
-	failedChange := false
-	for _, changeCH := range change.CustomHostnames {
-		if recordTypeCustomHostnameSupported[string(change.ResourceRecord.Type)] && changeCH.hostname != "" {
-			log.WithFields(logFields).Infof("Creating custom hostname %q", changeCH.hostname)
-			if ch, err := getcustomHostname(chs, changeCH.hostname); err == nil {
-				if changeCH.customOriginServer == ch.customOriginServer {
-					log.WithFields(logFields).Warnf("custom hostname %q already exists with the same origin %q, continue", changeCH.hostname, ch.customOriginServer)
-				} else {
-					failedChange = true
-					log.WithFields(logFields).Errorf("failed to create custom hostname, %q already exists with origin %q", changeCH.hostname, ch.customOriginServer)
-				}
-			} else {
-				chErr := p.Client.CreatecustomHostname(ctx, zoneID, changeCH)
-				if chErr != nil {
-					failedChange = true
-					log.WithFields(logFields).Errorf("failed to create custom hostname %q: %v", changeCH.hostname, chErr)
-				}
-			}
-		}
-	}
-	return !failedChange
-}
-
 // submitChanges takes a zone and a collection of Changes and sends them as a single transaction.
 func (p *CloudFlareProvider) submitChanges(ctx context.Context, changes []*cloudFlareChange) error {
 	// return early if there is nothing to change
@@ -909,24 +704,6 @@ func (p *CloudFlareProvider) getRecordID(records DNSRecordsMap, record dns.Recor
 	return ""
 }
 
-func getcustomHostname(chs customHostnamesMap, chName string) (customHostname, error) {
-	if chName == "" {
-		return customHostname{}, fmt.Errorf("failed to get custom hostname: %q is empty", chName)
-	}
-	if ch, ok := chs[customHostnameIndex{hostname: chName}]; ok {
-		return ch, nil
-	}
-	return customHostname{}, fmt.Errorf("failed to get custom hostname: %q not found", chName)
-}
-
-func (p *CloudFlareProvider) newcustomHostname(hostname string, origin string) customHostname {
-	return customHostname{
-		hostname:           hostname,
-		customOriginServer: origin,
-		ssl:                getCustomHostnamesSSLOptions(p.CustomHostnamesConfig),
-	}
-}
-
 func (p *CloudFlareProvider) newCloudFlareChange(action changeAction, ep *endpoint.Endpoint, target string, current *endpoint.Endpoint) (*cloudFlareChange, error) {
 	ttl := dns.TTL(defaultTTL)
 	proxied := shouldBeProxied(ep, p.proxiedByDefault)
@@ -942,7 +719,7 @@ func (p *CloudFlareProvider) newCloudFlareChange(action changeAction, ep *endpoi
 			prevCustomHostnames = getEndpointCustomHostnames(current)
 		}
 		for _, v := range getEndpointCustomHostnames(ep) {
-			newCustomHostnames[v] = p.newcustomHostname(v, ep.DNSName)
+			newCustomHostnames[v] = p.newCustomHostname(v, ep.DNSName)
 		}
 	}
 
@@ -1011,48 +788,6 @@ func (p *CloudFlareProvider) getDNSRecordsMap(ctx context.Context, zoneID string
 		return nil, convertCloudflareError(iter.Err())
 	}
 	return recordsMap, nil
-}
-
-func newcustomHostnameIndex(ch customHostname) customHostnameIndex {
-	return customHostnameIndex{hostname: ch.hostname}
-}
-
-// listCustomHostnamesWithPagination performs automatic pagination of results on requests to cloudflare.CustomHostnames
-func (p *CloudFlareProvider) listCustomHostnamesWithPagination(ctx context.Context, zoneID string) (customHostnamesMap, error) {
-	if !p.CustomHostnamesConfig.Enabled {
-		return nil, nil
-	}
-	chs := make(customHostnamesMap)
-	iter := p.Client.CustomHostnames(ctx, zoneID)
-	customHostnames, err := listAllCustomHostnames(iter)
-	if err != nil {
-		convertedError := convertCloudflareError(err)
-		if !errors.Is(convertedError, provider.SoftError) {
-			log.Errorf("zone %q failed to fetch custom hostnames. Please check if \"Cloudflare for SaaS\" is enabled and API key permissions, %v", zoneID, err)
-		}
-		return nil, convertedError
-	}
-	for _, ch := range customHostnames {
-		chs[newcustomHostnameIndex(ch)] = ch
-	}
-	return chs, nil
-}
-
-func getCustomHostnamesSSLOptions(customHostnamesConfig CustomHostnamesConfig) *customHostnameSSL {
-	ssl := &customHostnameSSL{
-		sslType:      "dv",
-		method:       "http",
-		bundleMethod: "ubiquitous",
-		settings: customHostnameSSLSettings{
-			minTLSVersion: customHostnamesConfig.MinTLSVersion,
-		},
-	}
-	// Set CertificateAuthority if provided
-	// We're not able to set it at all (even with a blank) if you're not on an enterprise plan
-	if customHostnamesConfig.CertificateAuthority != "none" {
-		ssl.certificateAuthority = customHostnamesConfig.CertificateAuthority
-	}
-	return ssl
 }
 
 func shouldBeProxied(ep *endpoint.Endpoint, proxiedByDefault bool) bool {
